task1-serverless-inference/requirements.txt
azure-ai-inference>=1.0.0b9
python-dotenv>=1.0.0
Pillow>=10.0.0
